requirements.txt
attrs
certifi
chardet
coverage
future
idna
importlib-metadata
nba-api
numpy
oauthlib
packaging
panda
pandas
pluggy
py
pyparsing
pytest
pytest-cov
python-dateutil
python-twitter
pytz
requests
requests-oauthlib
schedule
six
urllib3
wcwidth
zipp
